refactor: rename file for consistency
